api endpoints

+++ pages/api/products/[id].js
@@ -5,8 +5,10 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
-
+  const token = cookies.token;
+  
   // Connect to dbConnect
   await dbConnect();
 
@@ -15,21 +17,27 @@ export default async function handler(req, res) {
       const product = await Product.findById(id);
       res.status(200).json(product);
     } catch (err) {
-      res.status.json({ message: err.message });
+      res.status(500).json({ message: err.message });
     }
   }
   if (method === "PUT") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not authenticated!");
+    }
     // update product
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
       }); //new:true returns the most updated version
-      res.status(201).json(product);
+      res.status(200).json(product);
     } catch (err) {
       res.status(500).json({ message: err.message });
     }
   }
   if (method === "DELETE") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not authenticated!");
+    }
     // delete product
     try {
       await Product.findByIdAndDelete(id);
